backend/requirements.txt
APScheduler==3.11.2
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.1
numpy==2.4.1
pandas==3.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-magic-bin==0.4.14
pytz==2025.2
reportlab==4.4.9
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
